requirements.txt
google-genai
python-dotenv
psycopg2-binary
PyPDF2
pgvector